lessons/276/fastapi-app-simpler/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi[standard]==0.127.1
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
granian[pname]==2.6.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.5
packaging==25.0
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.21
PyYAML==6.0.3
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
sentry-sdk==2.48.0
setuptools==80.9.0
shellingham==1.5.4
socketify==0.0.31
starlette==0.50.0
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid6==2025.0.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
